+++ lib/circuit_switch/engine.rb
@@ -1,0 +1,9 @@
+module CircuitSwitch
+  def self.table_name_prefix
+    ''
+  end
+
+  class Engine < ::Rails::Engine
+    isolate_namespace ::CircuitSwitch
+  end
+end